Move this down to IPSocket.
